# docs/openapi/adept-api-v1.json
{
  "components" : {
    "schemas" : {
...
      "AuthSessionResponse" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/AuthenticatedSessionResponse"
        }, {
          "$ref" : "#/components/schemas/WorkspaceSelectionSessionResponse"
        } ]
      },
...
      "EmailRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "FieldError" : {
        "additionalProperties" : false,
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ],
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string",
            "writeOnly" : true
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "additionalProperties" : false,
        "description" : "RFC 9457 problem response with stable Adept fields.",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "fieldErrors" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            },
            "type" : "array"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "code", "detail", "instance", "status", "title", "traceId", "type" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "HTTP Bearer JWT access token.",
        "scheme" : "bearer",
        "type" : "http"
      },
      "csrfHeader" : {
        "description" : "CSRF header whose value must match the paired readable XSRF-TOKEN cookie.",
        "in" : "header",
        "name" : "X-XSRF-TOKEN",
        "type" : "apiKey"
      },
      "oauthSessionCookie" : {
        "description" : "Short-lived HttpOnly session used for Google login, onboarding, or reauthentication.",
        "in" : "cookie",
        "name" : "adept_oauth",
        "type" : "apiKey"
      },
      "refreshCookie" : {
        "description" : "HttpOnly refresh-session cookie. The cookie is required for refresh, workspace switching, and zero-workspace account recovery.",
        "in" : "cookie",
        "name" : "adept_refresh",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "Adept Industries Phase 2 API contract.",
    "title" : "Adept API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/forgot-password" : {
      "post" : {
        "operationId" : "forgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Request accepted",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Validation failed or the request was malformed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "CSRF, origin, membership, or role authorization failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "413" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body exceeded the 16 KiB limit",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "415" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body used an unsupported media type",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "A request rate limit was exceeded",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Retry-After" : {
                "description" : "Seconds until the request may be retried.",
                "schema" : {
                  "format" : "int64",
                  "minimum" : 0,
                  "type" : "integer"
                }
              }
            }
          }
        },
        "security" : [ {
          "csrfHeader" : [ ]
        } ],
        "summary" : "Request a password reset",
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description" : "Session created",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Validation failed or the request was malformed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Authentication or session validation failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "CSRF, origin, membership, or role authorization failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "413" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body exceeded the 16 KiB limit",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "415" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body used an unsupported media type",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "A request rate limit was exceeded",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Retry-After" : {
                "description" : "Seconds until the request may be retried.",
                "schema" : {
                  "format" : "int64",
                  "minimum" : 0,
                  "type" : "integer"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "csrfHeader" : [ ]
        } ],
        "summary" : "Create a browser session",
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "description" : "Idempotently clears session cookies. A valid adept_refresh cookie is revoked when present; a missing cookie still returns 204.",
        "operationId" : "logoutSession",
        "responses" : {
          "204" : {
            "description" : "Session cookies cleared",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "CSRF, origin, membership, or role authorization failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "A request rate limit was exceeded",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Retry-After" : {
                "description" : "Seconds until the request may be retried.",
                "schema" : {
                  "format" : "int64",
                  "minimum" : 0,
                  "type" : "integer"
                }
              },
              "Set-Cookie" : {
                "description" : "Sets or clears adept_refresh and rotates or expires XSRF-TOKEN as described by the operation.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "csrfHeader" : [ ]
        } ],
        "summary" : "End the browser session",
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/me" : {
      "get" : {
        "operationId" : "getCurrentUser",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "Current identity returned",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Authentication or session validation failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "CSRF, origin, membership, or role authorization failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Get the current identity context",
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/resend-verification" : {
      "post" : {
        "operationId" : "resendVerification",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Request accepted",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Validation failed or the request was malformed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "CSRF, origin, membership, or role authorization failed",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "413" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body exceeded the 16 KiB limit",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "415" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body used an unsupported media type",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              }
            }
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "A request rate limit was exceeded",
            "headers" : {
              "Cache-Control" : {
                "description" : "Sensitive Phase 2 responses are not cached.",
                "schema" : {
                  "enum" : [ "no-store" ],
                  "type" : "string"
                }
              },
              "Retry-After" : {
                "description" : "Seconds until the request may be retried.",
                "schema" : {
                  "format" : "int64",
                  "minimum" : 0,
                  "type" : "integer"
                }
              }
            }
          }
        },
        "security" : [ {
          "csrfHeader" : [ ]
        } ],
        "summary" : "Request another verification email",
        "tags" : [ "auth-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}